fix: [DPE-6234] Disable general_log and plugin audit_log_filter
The config option general_log is not compatible with the real
productions due to dumping all SQL queries to the general.log:
https://dev.mysql.com/doc/refman/8.0/en/query-log.html

> ... The server writes information to this log when clients connect or
> disconnect, and it logs each SQL statement received from clients...

In the past, it was a necessary evil as we had to keep record all
connects/disconnects from the clients for security reasons.
Nowadays we have audit_log plugin enabled by default, therefor no need to
pay such a high performance price and struggle from the general.log size

https://github.com/canonical/mysql-operator/issues/523#issuecomment-2540596167
>  33GB /var/snap/charmed-mysql/common/var/log/mysql/archive_general/general.log-20241212_0013

Additionally, we are working on the logs compression spec (separate topic).

---

While working here, we have noticed unnecessary enabled plugin audit_log_filter
which also produces log files in /var/lib/mysql. Disabling it for now,
as it is not necessary for audit_log plugin announcer functionality.

We will consider re-enabling the plugin without rush in
https://warthogs.atlassian.net/browse/DPE-6235